api/requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-filter==2.1.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
oauthlib==3.0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.1
djangorestframework_simplejwt
django-cors-headers